fix(core): move get funtion to LottoNumbers for LottoNumbers
