+++ src/main/kotlin/study/lotto/controller/LottoController.kt
@@ -3,6 +3,7 @@ package study.lotto.controller
 import study.lotto.domain.Lotto
 import study.lotto.domain.LottoGameResult
 import study.lotto.domain.LottoNumber
+import study.lotto.domain.LottoNumbers
 import study.lotto.domain.Lottoes
 import study.lotto.view.InputView
 import study.lotto.view.ResultView
@@ -16,7 +17,8 @@ class LottoController(
         resultView.displayLottoes(buyingLottoes)
 
         val lastWeekWinningLotto = inputView.getLastWeekWinningNumbers()
-            .let(Lotto::get)
+            .let(LottoNumbers::get)
+            .let(::Lotto)
 
         val bonusNumber = inputView.getBonusNumber().let(::LottoNumber)
