+++ src/test/kotlin/study/lotto/domain/LottoesTest.kt
@@ -13,7 +13,7 @@ class LottoesTest {
 
     @Test
     fun `getPrizes 함수가 정상적으로 일치 등급을 반환해야 한다`() {
-        val winningLotto = Lotto.get(listOf(1, 2, 3, 4, 5, 6))
+        val winningLotto = Lotto(LottoNumbers.get(listOf(1, 2, 3, 4, 5, 6)))
         val bonusNumber = LottoNumber(10)
         val lottos = Lottoes(listOf(winningLotto))
         val expectedGrade = listOf(PrizeGrade.GRADE_1)
